+++ docs/changelog.md
@@ -8,7 +8,7 @@ New:
 
 * Finite bandwidth correction: optionally account for the effect of the laser pulse duration when sampling photon momenta.
 * Electron-positron pair creation, via LMA and LCFA rates. `leading-order-only` feature, if enabled, replaces the LMA rate with a perturbative equivalent.
-* Previous behaviour, i.e. no pair creation or photon tracking, is recovered by compiling with `no-pair-creation` feature.
+* Previous behaviour, i.e. no pair creation or photon tracking, is recovered by running with `pair_creation: false` in the input file.
 * Photon primaries, with either Gaussian or bremsstrahlung energy spectra.
 * Automatic naming of output files.